Create getStatus() function, use it in initialize
First step in separating the the status() method that does two things.
status() is not too functional as it tries to be two things at the
same time. It's a setter and a getter depending on the action parameter.
Not to clean-codey so it needs to be separated. getStatus() is the
initial step towards that.